Try to speed up rendering of data views
It turns out that updating a tree view in Qt may be very slow. This
branch contains some ideas to speed up updating of data object views.
One idea is to use a QTableWidget instead of a QTreeView. This could
potentially be much faster, but at the cost of more complicated code,
and not so pretty view.

Some notes. It turns out that one source of slowdown is:
ui->treeViewDataObjects->header()->setSectionResizeMode(0, QHeaderView::ResizeToContents);

This can be *VERY* slow. Don't use it.

Also, this line speeds things up quite a bit:
ui->treeViewDataObjects->setUniformRowHeights(true);